+++ src/exportProvider.ts
@@ -1,11 +1,10 @@
 'use strict';
 
 import * as vscode from 'vscode';
-
 import * as path from 'path';
 import * as fs from 'fs';
 
-import { mjml2html } from 'mjml';
+import helper from './helper';
 
 export default class ExportHTML {
 
@@ -18,45 +17,42 @@ export default class ExportHTML {
 	}
 
     private export() {
-		if (!(vscode.window.activeTextEditor.document.languageId === 'mjml')) {
+		if (!(helper.isMJMLFile(vscode.window.activeTextEditor.document))) {
 			vscode.window.showWarningMessage('This is not a MJML document!');
 			return;
 		}
 
-		vscode.window.showInputBox({ placeHolder: 'File name' }).then(
-			(fileName) => {
+		let defaultFileName = path.basename(vscode.window.activeTextEditor.document.uri.fsPath).replace(/\.[^\.]+$/, '');
+
+		vscode.window.showInputBox({ placeHolder: 'File name (' + defaultFileName + '.html)' }).then((fileName) => {
+			let content = helper.renderMJML(
+				vscode.window.activeTextEditor.document.getText(),
+				vscode.workspace.getConfiguration('mjml').minifyHtmlOutput,
+				vscode.workspace.getConfiguration('mjml').beautifyHtmlOutput
+			);
+
+			if (content) {
 				if (fileName) {
 					fileName = fileName.replace(/\.[^\.]+$/, '');
 				}
 				else {
-					fileName = path.basename(vscode.window.activeTextEditor.document.uri.fsPath).replace(/\.[^\.]+$/, '');
+					fileName = defaultFileName;
 				}
 
 				let file = path.resolve(vscode.window.activeTextEditor.document.uri.fsPath, '../' + fileName + '.html');
-				try {
-					let content = mjml2html(vscode.window.activeTextEditor.document.getText(), {
-						level: 'skip',
-						filePath: vscode.window.activeTextEditor.document.uri.fsPath
-					});
 
-					if (content.html) {
-						fs.writeFile(file, content.html, function (err) {
-							if (err) {
-								return console.log(err);
-							}
+				fs.writeFile(file, content, function (err) {
+					if (err) {
+						return console.log(err);
+					}
 
-							vscode.window.showInformationMessage('File saved as ' + fileName + '.html');
-						});
-					}
-					else {
-						vscode.window.showErrorMessage('MJMLError: Failed to parse file ' + path.basename(vscode.window.activeTextEditor.document.uri.fsPath));
-					}
-				}
-				catch (e) {
-					vscode.window.showErrorMessage('MJMLError: Failed to parse file ' + path.basename(vscode.window.activeTextEditor.document.uri.fsPath));
-				}
+					vscode.window.showInformationMessage('File saved as ' + fileName + '.html');
+				});
 			}
-		);
+			else {
+				vscode.window.showErrorMessage('MJMLError: Failed to parse file ' + path.basename(vscode.window.activeTextEditor.document.uri.fsPath));
+			}
+		});
     }
 
 }